feat(core): WireHeader v2 마이그레이션 — 테스트 + 예제 + 벤치마크 전수 수정
- 모든 테스트 파일 msg_id uint16→uint32, flags uint16→uint8 반영
- 테스트 기대값: SIZE 10→12, CURRENT_VERSION 1→2, FrameSize 266→268
- BigEndianByteOrder 테스트 v2 레이아웃으로 재작성
- ParseV1Rejected, LargeMsgId(0xDEADBEEF) 테스트 추가
- 예제 서버(echo, chat, multicore_echo) 핸들러 시그니처 u32 반영
- 벤치마크(dispatcher, frame_pipeline) 핸들러 시그니처 u32 반영
- WireHeader 지정 초기화 필드 순서 v2에 맞게 정렬

+++ apex_core/benchmarks/bench_helpers.hpp
@@ -24,8 +24,8 @@ make_socket_pair(boost::asio::io_context& io) {
     return {std::move(client), std::move(server)};
 }
 
-/// Build a complete wire frame (10B header + payload).
-inline std::vector<uint8_t> build_frame(uint16_t msg_id, size_t payload_size) {
+/// Build a complete wire frame (12B header + payload).
+inline std::vector<uint8_t> build_frame(uint32_t msg_id, size_t payload_size) {
     std::vector<uint8_t> frame(apex::core::WireHeader::SIZE + payload_size);
 
     apex::core::WireHeader header;
