+++ apex_core/benchmarks/micro/bench_dispatcher.cpp
@@ -12,12 +12,12 @@ static void BM_Dispatcher_Lookup(benchmark::State& state) {
     boost::asio::io_context io_ctx;
     MessageDispatcher dispatcher;
     for (int i = 0; i < num_handlers; ++i) {
-        dispatcher.register_handler(static_cast<uint16_t>(i),
-            [](SessionPtr, uint16_t, std::span<const uint8_t>) -> boost::asio::awaitable<Result<void>> {
+        dispatcher.register_handler(static_cast<uint32_t>(i),
+            [](SessionPtr, uint32_t, std::span<const uint8_t>) -> boost::asio::awaitable<Result<void>> {
                 co_return ok();
             });
     }
-    uint16_t target_id = static_cast<uint16_t>(num_handlers / 2);
+    uint32_t target_id = static_cast<uint32_t>(num_handlers / 2);
     for (auto _ : state) {
         bool has = dispatcher.has_handler(target_id);
         benchmark::DoNotOptimize(has);
